+++ src/components/Options.js
@@ -1,0 +1,22 @@
+import React from "react";
+import Option from "./Option";
+
+const Options = (props) => {
+    return (
+        <div id='options'>
+            <button onClick={props.handleDeleteOptions}>Remove All</button>
+            {props.options.length ===0 && <p>Please add an option to get started...</p>}
+            {
+                props.options.map(option => (
+                    <Option
+                        key={option}
+                        optionText={option}
+                        handleDeleteOption={props.handleDeleteOption}
+                    />
+                ))
+            }
+        </div>
+    );
+}
+
+export default Options;
